--- project/src/main/java/ua/ies/tut1/RabbitAmpqTutorialsApplication.java
@@ -1,32 +1,0 @@
-package ua.ies.tut1;
-
-import org.springframework.boot.CommandLineRunner;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Profile;
-import org.springframework.scheduling.annotation.EnableScheduling;
-
-
-@SpringBootApplication
-@EnableScheduling
-public class RabbitAmpqTutorialsApplication {
-    @Profile("usage_message")
-    @Bean
-    public CommandLineRunner usage() {
-        return args -> {
-            System.out.println("This app uses Spring Profiles to "+
-                "control its behaviour. (and i dont understand anything.\n");
-        };
-    }
-
-    @Profile("!usage_message")
-    @Bean
-    public CommandLineRunner tutorial() {
-        return new RabbitAmqpTutorialsRunner();
-    }
-
-    public static void main(String[] args) {
-        SpringApplication.run(RabbitAmpqTutorialsApplication.class, args);
-    }
-}
