--- project/src/main/java/ua/ies/tut1/Tut1Config.java
@@ -1,28 +1,0 @@
-package ua.ies.tut1;
-
-
-import org.springframework.amqp.core.Queue;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Profile;
-
-@Profile({"tut1", "hello-world"})
-@Configuration
-public class Tut1Config {
-    @Bean
-    public Queue co2() {
-        return new Queue("co2", false, false, false);
-    }
-
-    @Profile("receiver")
-    @Bean
-    public Tut1Receiver receiver() {
-        return new Tut1Receiver();
-    }
-
-    @Profile("sender")
-    @Bean
-    public Tut1Sender sender() {
-        return new Tut1Sender();
-    }
-}
