--- project/src/main/java/ua/ies/tut1/Tut1Receiver.java
@@ -1,13 +1,0 @@
-package ua.ies.tut1;
-
-import org.springframework.amqp.rabbit.annotation.RabbitHandler;
-import org.springframework.amqp.rabbit.annotation.RabbitListener;
-
-@RabbitListener(queues = "co2")
-public class Tut1Receiver {
-
-    @RabbitHandler
-    public void receive(String in) {
-        System.out.println(" [x] Received '" + in + "'");
-    }
-}
